requirements.txt
numpy>=1.22
scipy>=1.10
pandas>=1.5
pyyaml>=6.0
torch>=2.0
pulp>=2.7
tqdm>=4.66
matplotlib>=3.7
